+++ examples/basic/android/app/src/main/java/com/videoplayer/MainApplication.java
@@ -2,6 +2,7 @@ package com.videoplayer;
 
 import android.app.Application;
 
+import androidx.multidex.MultiDexApplication;
 import com.facebook.react.ReactApplication;
 import com.brentvatne.react.ReactVideoPackage;
 import com.facebook.react.ReactNativeHost;
@@ -9,10 +10,12 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
 
+import com.reactnativecommunity.picker.RNCPickerPackage;
+
 import java.util.Arrays;
 import java.util.List;
 
-public class MainApplication extends Application implements ReactApplication {
+public class MainApplication extends MultiDexApplication implements ReactApplication {
 
   private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
     @Override
@@ -24,8 +27,14 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.asList(
         new MainReactPackage(),
-        new ReactVideoPackage()
+        new ReactVideoPackage(),
+        new RNCPickerPackage()
       );
+    }
+
+    @Override
+    protected String getJSMainModuleName() {
+      return "src/index";
     }
   };