+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Better support newer versions of RNW (64 and newer) [#2535](https://github.com/react-native-video/react-native-video/pull/2535)
 - Fix nil string uri parameter error [#695](https://github.com/react-native-video/react-native-video/pull/695)
 - (Breaking) Bump shaka-player to 3.3.2 [#2587](https://github.com/react-native-video/react-native-video/pull/2587)
+- Improve basic player example on android [#2662](https://github.com/react-native-video/react-native-video/pull/2662)
 
 ### Version 5.2.0